Add a few more extensions for pygments, and fix markdown
Markdown is known as `md` to pygments, according to `pygments -L`.
And our ssr snapshots are js code, so we can format them like that.

This change doesn't actually do anything, since this config file is
overriden by the database entry at
   https://phabricator.khanacademy.org/config/edit/syntax.filemap/
But I feel it's good to have a record in source control.

Auditors: benkraft

Test plan:
Made the chnages in the database listed above